add protection around schema changes

+++ spec/sql/sharded.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `sharded` (
-  `id` int(10) NOT NULL AUTO_INCREMENT,
+  `id` bigint(20) NOT NULL AUTO_INCREMENT,
   `account_id` int(11) NOT NULL,
   `nice_id` int(11) NOT NULL,
   `status_id` tinyint NOT NULL default 2,
